+++ tools/analyzer/index.cpp
@@ -28,8 +28,8 @@ namespace arsh::lsp {
 
 void DeclBase::addRef(SymbolRef ref) {
   auto iter = std::lower_bound(this->refs.begin(), this->refs.end(), ref);
-  if (iter != this->refs.end() && (*iter).getModId() == ref.getModId() &&
-      (*iter).getPos() == ref.getPos()) {
+  if (iter != this->refs.end() && iter->getModId() == ref.getModId() &&
+      iter->getPos() == ref.getPos()) {
     *iter = ref; // update
   } else {
     this->refs.insert(iter, ref);
@@ -81,6 +81,7 @@ std::string DeclSymbol::mangle(StringRef recvTypeName, Kind k, StringRef name) {
   case Kind::MOD:
   case Kind::MOD_CONST:
   case Kind::PARAM:
+  case Kind::GENERIC_METHOD_PARAM:
   case Kind::HERE_START:
     value = name.toString();
     break;
@@ -130,7 +131,8 @@ std::pair<StringRef, StringRef> DeclSymbol::demangleWithRecv(Kind k, Attr a,
   case Kind::MOD_CONST:
   case Kind::HERE_START:
     break;
-  case Kind::PARAM: {
+  case Kind::PARAM:
+  case Kind::GENERIC_METHOD_PARAM: {
     auto r = mangledName.find('+');
     assert(r != StringRef::npos);
     mangledName = mangledName.substr(0, r);